refactor: move queries to context modules +  move common logic into new ActivityHelpers
