+++ lib/oli_web/components/delivery/practice_activities/practice_activities.ex
@@ -1,34 +1,13 @@
 defmodule OliWeb.Components.Delivery.PracticeActivities do
   use OliWeb, :live_component
 
-  import Ecto.Query
-
-  alias Oli.Accounts.User
-
-  alias Oli.Analytics.Summary.{
-    ResourcePartResponse,
-    ResourceSummary,
-    ResponseSummary,
-    StudentResponse
-  }
-
   alias OliWeb.Common.InstructorDashboardPagedTable
   alias Oli.Delivery.Attempts.Core
 
-  alias Oli.Delivery.Attempts.Core.{
-    ActivityAttempt,
-    ResourceAccess,
-    ResourceAttempt
-  }
-
   alias Oli.Delivery.Sections
-  alias Oli.Delivery.Sections.Section
-  alias Oli.Publishing.DeliveryResolver
-  alias Oli.Repo
-  alias Oli.Resources.Revision
-  alias Oli.Resources.ResourceType
   alias OliWeb.Common.{Params, SearchInput}
   alias OliWeb.Common.Table.SortableTableModel
+  alias OliWeb.Delivery.ActivityHelpers
   alias OliWeb.Delivery.PracticeActivities.PracticeAssessmentsTableModel
   alias OliWeb.ManualGrading.RenderedActivity
   alias OliWeb.Router.Helpers, as: Routes
@@ -408,9 +387,9 @@ defmodule OliWeb.Components.Delivery.PracticeActivities do
       end)
 
     current_activities =
-      get_activities(
-        current_assessment,
-        section,
+      ActivityHelpers.get_activities(
+        current_assessment.resource_id,
+        section.id,
         student_ids
       )
 
@@ -418,7 +397,7 @@ defmodule OliWeb.Components.Delivery.PracticeActivities do
       Enum.map(current_activities, fn activity -> activity.resource_id end)
 
     activities_details =
-      get_activities_details(
+      ActivityHelpers.get_activities_details(
         activity_resource_ids,
         socket.assigns.section,
         socket.assigns.activity_types_map,
@@ -435,7 +414,7 @@ defmodule OliWeb.Components.Delivery.PracticeActivities do
         activity
         |> Map.put(:preview_rendered, get_preview_rendered(activity_details, socket))
         |> Map.put(:datasets, Map.get(activity_details, :datasets))
-        |> add_activity_attempts_info(students, student_ids, section)
+        |> ActivityHelpers.add_activity_attempts_info(students, student_ids, section)
       end)
 
     {:noreply,
@@ -679,29 +658,6 @@ defmodule OliWeb.Components.Delivery.PracticeActivities do
     )
   end
 
-  defp add_activity_attempts_info(activity, students, student_ids, section) do
-    students_with_attempts =
-      DeliveryResolver.students_with_attempts_for_page(
-        activity,
-        section,
-        student_ids
-      )
-
-    student_emails_without_attempts =
-      Enum.reduce(students, [], fn s, acc ->
-        if s.id in students_with_attempts do
-          acc
-        else
-          [s.email | acc]
-        end
-      end)
-
-    activity
-    |> Map.put(:students_with_attempts_count, Enum.count(students_with_attempts))
-    |> Map.put(:student_emails_without_attempts, student_emails_without_attempts)
-    |> Map.put(:total_attempts_count, count_attempts(activity, section, student_ids) || 0)
-  end
-
   defp get_preview_rendered(nil, socket), do: socket
 
   defp get_preview_rendered(activity_attempt, socket) do
@@ -733,415 +689,6 @@ defmodule OliWeb.Components.Delivery.PracticeActivities do
       })
 
     assign(socket, table_model: table_model)
-  end
-
-  defp count_attempts(
-         current_activity,
-         %Section{analytics_version: :v2, id: section_id},
-         student_ids
-       ) do
-    page_type_id = ResourceType.get_id_by_type("activity")
-
-    from(rs in ResourceSummary,
-      where:
-        rs.section_id == ^section_id and rs.resource_id == ^current_activity.resource_id and
-          rs.user_id in ^student_ids and rs.project_id == -1 and rs.publication_id == -1 and
-          rs.resource_type_id == ^page_type_id,
-      select: sum(rs.num_attempts)
-    )
-    |> Repo.one()
-  end
-
-  defp count_attempts(current_activity, section, student_ids) do
-    from(ra in ResourceAttempt,
-      join: access in ResourceAccess,
-      on: access.id == ra.resource_access_id,
-      where:
-        ra.lifecycle_state == :evaluated and access.section_id == ^section.id and
-          access.resource_id == ^current_activity.resource_id and access.user_id in ^student_ids,
-      select: count(ra.id)
-    )
-    |> Repo.one()
-  end
-
-  defp get_activities(current_assessment, section, student_ids) do
-    from(aa in ActivityAttempt,
-      join: res_attempt in ResourceAttempt,
-      on: aa.resource_attempt_id == res_attempt.id,
-      where: aa.lifecycle_state == :evaluated,
-      join: res_access in ResourceAccess,
-      on: res_attempt.resource_access_id == res_access.id,
-      where:
-        res_access.section_id == ^section.id and
-          res_access.resource_id == ^current_assessment.resource_id and
-          res_access.user_id in ^student_ids and is_nil(aa.survey_id),
-      join: rev in Revision,
-      on: aa.revision_id == rev.id,
-      group_by: [rev.resource_id, rev.id],
-      select: map(rev, [:id, :resource_id, :title])
-    )
-    |> Repo.all()
-  end
-
-  def get_activities_details(activity_resource_ids, section, activity_types_map, page_resource_id) do
-    multiple_choice_type_id =
-      Enum.find_value(activity_types_map, fn {k, v} -> if v.title == "Multiple Choice", do: k end)
-
-    single_response_type_id =
-      Enum.find_value(activity_types_map, fn {k, v} -> if v.title == "Single Response", do: k end)
-
-    multi_input_type_id =
-      Enum.find_value(activity_types_map, fn {k, v} ->
-        if v.title == "Multi Input",
-          do: k
-      end)
-
-    likert_type_id =
-      Enum.find_value(activity_types_map, fn {k, v} -> if v.title == "Likert", do: k end)
-
-    activity_attempts =
-      from(activity_attempt in ActivityAttempt,
-        left_join: resource_attempt in assoc(activity_attempt, :resource_attempt),
-        left_join: resource_access in assoc(resource_attempt, :resource_access),
-        left_join: user in assoc(resource_access, :user),
-        left_join: activity_revision in assoc(activity_attempt, :revision),
-        left_join: resource_revision in assoc(resource_attempt, :revision),
-        where:
-          resource_access.section_id == ^section.id and
-            activity_revision.resource_id in ^activity_resource_ids,
-        select: activity_attempt,
-        select_merge: %{
-          activity_type_id: activity_revision.activity_type_id,
-          activity_title: activity_revision.title,
-          page_title: resource_revision.title,
-          page_id: resource_revision.resource_id,
-          resource_attempt_number: resource_attempt.attempt_number,
-          graded: resource_revision.graded,
-          user: user,
-          revision: activity_revision,
-          resource_attempt_guid: resource_attempt.attempt_guid,
-          resource_access_id: resource_access.id
-        }
-      )
-      |> Repo.all()
-
-    if section.analytics_version == :v2 do
-      response_summaries =
-        from(rs in ResponseSummary,
-          join: rpp in ResourcePartResponse,
-          on: rs.resource_part_response_id == rpp.id,
-          left_join: sr in StudentResponse,
-          on:
-            rs.section_id == sr.section_id and rs.page_id == sr.page_id and
-              rs.resource_part_response_id == sr.resource_part_response_id,
-          left_join: u in User,
-          on: sr.user_id == u.id,
-          where:
-            rs.section_id == ^section.id and rs.page_id == ^page_resource_id and
-              rs.publication_id == -1 and rs.project_id == -1 and
-              rs.activity_id in ^activity_resource_ids,
-          select: %{
-            part_id: rpp.part_id,
-            response: rpp.response,
-            count: rs.count,
-            user: u,
-            activity_id: rs.activity_id
-          }
-        )
-        |> Repo.all()
-
-      Enum.map(activity_attempts, fn activity_attempt ->
-        case activity_attempt.activity_type_id do
-          ^multiple_choice_type_id ->
-            add_choices_frequencies(activity_attempt, response_summaries)
-
-          ^single_response_type_id ->
-            add_single_response_details(activity_attempt, response_summaries)
-
-          ^multi_input_type_id ->
-            add_multi_input_details(activity_attempt, response_summaries)
-
-          ^likert_type_id ->
-            add_likert_details(activity_attempt, response_summaries)
-
-          _ ->
-            activity_attempt
-        end
-      end)
-    else
-      activity_attempts
-    end
-  end
-
-  defp add_single_response_details(activity_attempt, response_summaries) do
-    responses =
-      Enum.reduce(response_summaries, [], fn response_summary, acc ->
-        if response_summary.activity_id == activity_attempt.resource_id do
-          [
-            %{
-              text: response_summary.response,
-              user_name: OliWeb.Common.Utils.name(response_summary.user)
-            }
-            | acc
-          ]
-        else
-          acc
-        end
-      end)
-      |> Enum.reverse()
-
-    update_in(
-      activity_attempt,
-      [Access.key!(:revision), Access.key!(:content)],
-      &Map.put(&1, "responses", responses)
-    )
-  end
-
-  defp add_choices_frequencies(activity_attempt, response_summaries) do
-    responses =
-      Enum.filter(response_summaries, fn response_summary ->
-        response_summary.activity_id == activity_attempt.resource_id
-      end)
-
-    choices =
-      activity_attempt.transformed_model["choices"]
-      |> Enum.map(
-        &Map.merge(&1, %{
-          "frequency" =>
-            Enum.find(responses, %{count: 0}, fn r -> r.response == &1["id"] end).count
-        })
-      )
-      |> then(fn choices ->
-        blank_reponses = Enum.find(responses, fn r -> r.response == "" end)
-
-        if blank_reponses[:response] do
-          [
-            %{
-              "content" => [
-                %{
-                  "children" => [
-                    %{
-                      "text" =>
-                        "Blank attempt (user submitted assessment without selecting any choice for this activity)"
-                    }
-                  ],
-                  "type" => "p"
-                }
-              ],
-              "frequency" => blank_reponses.count
-            }
-            | choices
-          ]
-        else
-          choices
-        end
-      end)
-
-    update_in(
-      activity_attempt,
-      [Access.key!(:transformed_model)],
-      &Map.put(&1, "choices", choices)
-    )
-  end
-
-  defp add_multi_input_details(activity_attempt, response_summaries) do
-    mapper = build_input_mapper(activity_attempt.transformed_model["inputs"])
-
-    Enum.reduce(
-      activity_attempt.transformed_model["inputs"],
-      activity_attempt,
-      fn input, acc2 ->
-        case input["inputType"] do
-          response when response in ["numeric", "text"] ->
-            add_text_or_numeric_responses(
-              acc2,
-              response_summaries,
-              mapper
-            )
-
-          "dropdown" ->
-            add_dropdown_choices(acc2, response_summaries)
-        end
-      end
-    )
-  end
-
-  defp add_dropdown_choices(acc, response_summaries) do
-    add_choices_frequencies(acc, response_summaries)
-    |> update_in(
-      [
-        Access.key!(:transformed_model),
-        Access.key!("inputs"),
-        Access.filter(&(&1["inputType"] == "dropdown")),
-        Access.key!("choiceIds")
-      ],
-      &List.insert_at(&1, -1, "0")
-    )
-  end
-
-  defp add_text_or_numeric_responses(acumulator, response_summaries, mapper) do
-    responses =
-      relevant_responses(acumulator.resource_id, response_summaries, mapper)
-
-    update_in(
-      acumulator,
-      [Access.key!(:transformed_model), Access.key!("authoring")],
-      &Map.put(&1, "responses", responses)
-    )
-  end
-
-  defp relevant_responses(resource_id, response_summaries, mapper) do
-    Enum.reduce(response_summaries, [], fn response_summary, acc_responses ->
-      if response_summary.activity_id == resource_id do
-        [
-          %{
-            text: response_summary.response,
-            user_name: OliWeb.Common.Utils.name(response_summary.user),
-            type: mapper[response_summary.part_id],
-            part_id: response_summary.part_id
-          }
-          | acc_responses
-        ]
-      else
-        acc_responses
-      end
-    end)
-  end
-
-  defp build_input_mapper(inputs) do
-    Enum.into(inputs, %{}, fn input ->
-      {input["partId"], input["inputType"]}
-    end)
-  end
-
-  defp add_likert_details(activity_attempt, response_summaries) do
-    {ordered_questions, question_text_mapper} =
-      Enum.reduce(
-        activity_attempt.revision.content["items"],
-        %{ordered_questions: [], question_text_mapper: %{}},
-        fn q, acc ->
-          question = %{
-            id: q["id"],
-            text: q["content"] |> hd() |> Map.get("children") |> hd() |> Map.get("text")
-          }
-
-          %{
-            ordered_questions: [question | acc.ordered_questions],
-            question_text_mapper: Map.put(acc.question_text_mapper, q["id"], question.text)
-          }
-        end
-      )
-      |> then(fn acc ->
-        {Enum.reverse(acc.ordered_questions), acc.question_text_mapper}
-      end)
-
-    {ordered_choices, choice_mapper} =
-      Enum.reduce(
-        activity_attempt.revision.content["choices"],
-        %{ordered_choices: [], choice_mapper: %{}, aux_points: 1},
-        fn ch, acc ->
-          choice = %{
-            id: ch["id"],
-            text: ch["content"] |> hd() |> Map.get("children") |> hd() |> Map.get("text"),
-            points: acc.aux_points
-          }
-
-          %{
-            ordered_choices: [choice | acc.ordered_choices],
-            choice_mapper:
-              Map.put(acc.choice_mapper, ch["id"], %{text: choice.text, points: choice.points}),
-            aux_points: acc.aux_points + 1
-          }
-        end
-      )
-      |> then(fn acc ->
-        {Enum.reverse(acc.ordered_choices), acc.choice_mapper}
-      end)
-
-    responses =
-      Enum.reduce(response_summaries, [], fn response_summary, acc ->
-        if response_summary.activity_id == activity_attempt.resource_id do
-          [
-            %{
-              count: response_summary.count,
-              choice_id: response_summary.response,
-              selected_choice_text:
-                Map.get(choice_mapper, to_string(response_summary.response))[:text],
-              selected_choice_points:
-                Map.get(choice_mapper, to_string(response_summary.response))[:points],
-              question_id: response_summary.part_id,
-              question: Map.get(question_text_mapper, to_string(response_summary.part_id))
-            }
-            | acc
-          ]
-        else
-          acc
-        end
-      end)
-
-    {average_points_per_question_id, responses_per_question_id} =
-      Enum.reduce(responses, {%{}, %{}}, fn response, {avg_points_acc, responses_acc} ->
-        {Map.put(avg_points_acc, response.question_id, [
-           response.selected_choice_points | Map.get(avg_points_acc, response.question_id, [])
-         ]),
-         Map.put(
-           responses_acc,
-           response.question_id,
-           Map.get(responses_acc, response.question_id, 0) + 1
-         )}
-      end)
-      |> then(fn {points_per_question_id, responses_per_question_id} ->
-        average_points_per_question_id =
-          Enum.into(points_per_question_id, %{}, fn {question_id, points} ->
-            count = Enum.count(points)
-
-            {
-              question_id,
-              if count == 0 do
-                0
-              else
-                Enum.sum(points) / count
-              end
-            }
-          end)
-
-        {average_points_per_question_id, responses_per_question_id}
-      end)
-
-    first_choice_text = Enum.at(ordered_choices, 0)[:text]
-    last_choice_text = Enum.at(ordered_choices, -1)[:text]
-
-    medians =
-      Enum.map(ordered_questions, fn q ->
-        %{
-          question: q.text,
-          median: Map.get(average_points_per_question_id, q.id, 0.0),
-          lo: first_choice_text,
-          hi: last_choice_text
-        }
-      end)
-
-    values =
-      Enum.map(responses, fn r ->
-        %{
-          value: r.selected_choice_points,
-          choice: r.selected_choice_text,
-          question: r.question,
-          out_of: Map.get(responses_per_question_id, r.question_id, 0)
-        }
-      end)
-
-    Map.merge(activity_attempt, %{
-      datasets: %{
-        medians: medians,
-        values: values,
-        questions_count: length(ordered_questions),
-        axis_values: Enum.map(ordered_choices, fn c -> c.points end),
-        first_choice_text: first_choice_text,
-        last_choice_text: last_choice_text,
-        title: activity_attempt.revision.title
-      }
-    })
   end
 
   defp build_units_and_modules(container_count, modules_and_units) do
