+++ lib/oli_web/components/delivery/surveys/surveys.ex
@@ -1,23 +1,8 @@
 defmodule OliWeb.Components.Delivery.Surveys do
   use OliWeb, :live_component
 
-  import Ecto.Query
-  alias Oli.Accounts.User
-
-  alias Oli.Analytics.Summary.ResourcePartResponse
-  alias Oli.Analytics.Summary.ResourceSummary
-  alias Oli.Analytics.Summary.ResponseSummary
-  alias Oli.Analytics.Summary.StudentResponse
+  alias OliWeb.Delivery.ActivityHelpers
   alias Oli.Delivery.Attempts.Core
-  alias Oli.Delivery.Attempts.Core.ActivityAttempt
-  alias Oli.Delivery.Attempts.Core.ResourceAccess
-  alias Oli.Delivery.Attempts.Core.ResourceAttempt
-  alias Oli.Delivery.Sections.Section
-  alias Oli.Publishing.DeliveryResolver
-  alias Oli.Repo
-  alias Oli.Resources.ResourceType
-  alias Oli.Resources.Revision
-
   alias OliWeb.Common.InstructorDashboardPagedTable
   alias OliWeb.Common.Params
   alias OliWeb.Common.SearchInput
@@ -322,13 +307,19 @@ defmodule OliWeb.Components.Delivery.Surveys do
   end
 
   defp find_current_activities(current_assessment, section, student_ids, students, socket) do
-    activities = get_activities(current_assessment, section, student_ids)
+    activities =
+      ActivityHelpers.get_activities(
+        current_assessment.resource_id,
+        section.id,
+        student_ids,
+        true
+      )
 
     activity_resource_ids =
       Enum.map(activities, fn activity -> activity.resource_id end)
 
     activities_details =
-      get_activities_details(
+      ActivityHelpers.get_activities_details(
         activity_resource_ids,
         socket.assigns.section,
         socket.assigns.activity_types_map,
@@ -342,7 +333,7 @@ defmodule OliWeb.Components.Delivery.Surveys do
         end)
 
       Map.put(activity, :preview_rendered, get_preview_rendered(activity_details, socket))
-      |> add_activity_attempts_info(students, student_ids, section)
+      |> ActivityHelpers.add_activity_attempts_info(students, student_ids, section)
     end)
   end
 
@@ -370,29 +361,6 @@ defmodule OliWeb.Components.Delivery.Surveys do
       })
 
     assign(socket, table_model: table_model)
-  end
-
-  defp add_activity_attempts_info(activity, students, student_ids, section) do
-    students_with_attempts =
-      DeliveryResolver.students_with_attempts_for_page(
-        activity,
-        section,
-        student_ids
-      )
-
-    student_emails_without_attempts =
-      Enum.reduce(students, [], fn s, acc ->
-        if s.id in students_with_attempts do
-          acc
-        else
-          [s.email | acc]
-        end
-      end)
-
-    activity
-    |> Map.put(:students_with_attempts_count, Enum.count(students_with_attempts))
-    |> Map.put(:student_emails_without_attempts, student_emails_without_attempts)
-    |> Map.put(:total_attempts_count, count_attempts(activity, section, student_ids) || 0)
   end
 
   defp get_preview_rendered(nil, socket), do: socket
@@ -519,365 +487,5 @@ defmodule OliWeb.Components.Delivery.Surveys do
       :surveys,
       params
     )
-  end
-
-  defp count_attempts(
-         current_activity,
-         %Section{analytics_version: :v2, id: section_id},
-         student_ids
-       ) do
-    page_type_id = ResourceType.get_id_by_type("activity")
-
-    from(rs in ResourceSummary,
-      where:
-        rs.section_id == ^section_id and rs.resource_id == ^current_activity.resource_id and
-          rs.user_id in ^student_ids and rs.project_id == -1 and rs.publication_id == -1 and
-          rs.resource_type_id == ^page_type_id,
-      select: sum(rs.num_attempts)
-    )
-    |> Repo.one()
-  end
-
-  defp count_attempts(current_activity, section, student_ids) do
-    from(ra in ResourceAttempt,
-      join: access in ResourceAccess,
-      on: access.id == ra.resource_access_id,
-      where:
-        ra.lifecycle_state == :evaluated and access.section_id == ^section.id and
-          access.resource_id == ^current_activity.resource_id and access.user_id in ^student_ids,
-      select: count(ra.id)
-    )
-    |> Repo.one()
-  end
-
-  def get_activities(current_assessment, section, student_ids) do
-    from(activity_attempt in ActivityAttempt,
-      join: resource_attempt in ResourceAttempt,
-      on: activity_attempt.resource_attempt_id == resource_attempt.id,
-      where: activity_attempt.lifecycle_state == :evaluated,
-      join: resource_accesses in ResourceAccess,
-      on: resource_attempt.resource_access_id == resource_accesses.id,
-      where:
-        resource_accesses.section_id == ^section.id and
-          resource_accesses.resource_id == ^current_assessment.resource_id and
-          resource_accesses.user_id in ^student_ids and not is_nil(activity_attempt.survey_id),
-      join: revision in Revision,
-      on: activity_attempt.revision_id == revision.id,
-      group_by: [revision.resource_id, revision.id],
-      select: map(revision, [:id, :resource_id, :title])
-    )
-    |> Repo.all()
-  end
-
-  @spec get_activities_details(
-          any(),
-          atom() | %{:analytics_version => any(), :id => any(), optional(any()) => any()},
-          any(),
-          any()
-        ) :: any()
-  def get_activities_details(activity_resource_ids, section, activity_types_map, page_resource_id) do
-    multiple_choice_type_id =
-      Enum.find_value(activity_types_map, fn {k, v} -> if v.title == "Multiple Choice", do: k end)
-
-    single_response_type_id =
-      Enum.find_value(activity_types_map, fn {k, v} -> if v.title == "Single Response", do: k end)
-
-    multi_input_type_id =
-      Enum.find_value(activity_types_map, fn {k, v} ->
-        if v.title == "Multi Input",
-          do: k
-      end)
-
-    likert_type_id =
-      Enum.find_value(activity_types_map, fn {k, v} -> if v.title == "Likert", do: k end)
-
-    activity_attempts =
-      from(activity_attempt in ActivityAttempt,
-        left_join: resource_attempt in assoc(activity_attempt, :resource_attempt),
-        left_join: resource_access in assoc(resource_attempt, :resource_access),
-        left_join: user in assoc(resource_access, :user),
-        left_join: activity_revision in assoc(activity_attempt, :revision),
-        left_join: resource_revision in assoc(resource_attempt, :revision),
-        where:
-          resource_access.section_id == ^section.id and
-            activity_revision.resource_id in ^activity_resource_ids,
-        select: activity_attempt,
-        select_merge: %{
-          activity_type_id: activity_revision.activity_type_id,
-          activity_title: activity_revision.title,
-          page_title: resource_revision.title,
-          page_id: resource_revision.resource_id,
-          resource_attempt_number: resource_attempt.attempt_number,
-          graded: resource_revision.graded,
-          user: user,
-          revision: activity_revision,
-          resource_attempt_guid: resource_attempt.attempt_guid,
-          resource_access_id: resource_access.id
-        }
-      )
-      |> Repo.all()
-
-    if section.analytics_version == :v2 do
-      response_summaries =
-        from(rs in ResponseSummary,
-          join: rpp in ResourcePartResponse,
-          on: rs.resource_part_response_id == rpp.id,
-          left_join: sr in StudentResponse,
-          on:
-            rs.section_id == sr.section_id and rs.page_id == sr.page_id and
-              rs.resource_part_response_id == sr.resource_part_response_id,
-          left_join: u in User,
-          on: sr.user_id == u.id,
-          where:
-            rs.section_id == ^section.id and rs.page_id == ^page_resource_id and
-              rs.publication_id == -1 and rs.project_id == -1 and
-              rs.activity_id in ^activity_resource_ids,
-          select: %{
-            part_id: rpp.part_id,
-            response: rpp.response,
-            count: rs.count,
-            user: u,
-            activity_id: rs.activity_id
-          }
-        )
-        |> Repo.all()
-
-      Enum.map(activity_attempts, fn activity_attempt ->
-        case activity_attempt.activity_type_id do
-          ^multiple_choice_type_id ->
-            add_choices_frequencies(activity_attempt, response_summaries)
-
-          ^single_response_type_id ->
-            add_single_response_details(activity_attempt, response_summaries)
-
-          ^multi_input_type_id ->
-            add_multi_input_details(activity_attempt, response_summaries)
-
-          ^likert_type_id ->
-            add_likert_details(activity_attempt, response_summaries)
-
-          _ ->
-            activity_attempt
-        end
-      end)
-    else
-      activity_attempts
-    end
-  end
-
-  defp add_single_response_details(activity_attempt, response_summaries) do
-    responses =
-      Enum.reduce(response_summaries, [], fn response_summary, acc ->
-        if response_summary.activity_id == activity_attempt.resource_id do
-          [
-            %{
-              text: response_summary.response,
-              user_name: OliWeb.Common.Utils.name(response_summary.user)
-            }
-            | acc
-          ]
-        else
-          acc
-        end
-      end)
-      |> Enum.reverse()
-
-    update_in(
-      activity_attempt,
-      [Access.key!(:revision), Access.key!(:content)],
-      &Map.put(&1, "responses", responses)
-    )
-  end
-
-  defp add_choices_frequencies(activity_attempt, response_summaries) do
-    responses =
-      Enum.filter(response_summaries, fn response_summary ->
-        response_summary.activity_id == activity_attempt.resource_id
-      end)
-
-    # we must consider the case where a transformed model is present and if so, then use it
-    # otherwise, use the revision model. This block also returns a corresponding updater function
-    {model, updater} =
-      case activity_attempt.transformed_model do
-        nil ->
-          {activity_attempt.revision.content,
-           fn activity_attempt, choices ->
-             update_in(
-               activity_attempt,
-               [Access.key!(:revision), Access.key!(:content)],
-               &Map.put(&1, "choices", choices)
-             )
-           end}
-
-        transformed_model ->
-          {transformed_model,
-           fn activity_attempt, choices ->
-             update_in(
-               activity_attempt,
-               [Access.key!(:transformed_model)],
-               &Map.put(&1, "choices", choices)
-             )
-           end}
-      end
-
-    choices =
-      model["choices"]
-      |> Enum.map(
-        &Map.merge(&1, %{
-          "frequency" =>
-            Enum.find(responses, %{count: 0}, fn r -> r.response == &1["id"] end).count
-        })
-      )
-      |> then(fn choices ->
-        blank_reponses = Enum.find(responses, fn r -> r.response == "" end)
-
-        if blank_reponses[:response] do
-          [
-            %{
-              "content" => [
-                %{
-                  "children" => [
-                    %{
-                      "text" =>
-                        "Blank attempt (user submitted assessment without selecting any choice for this activity)"
-                    }
-                  ],
-                  "type" => "p"
-                }
-              ],
-              "frequency" => blank_reponses.count
-            }
-            | choices
-          ]
-        else
-          choices
-        end
-      end)
-
-    updater.(activity_attempt, choices)
-  end
-
-  defp add_likert_details(activity_attempt, response_summaries) do
-    responses =
-      Enum.filter(response_summaries, fn response_summary ->
-        response_summary.activity_id == activity_attempt.resource_id
-      end)
-
-    choices =
-      activity_attempt.revision.content["choices"]
-      |> Enum.map(
-        &Map.merge(&1, %{
-          "frequency" =>
-            Enum.find(responses, %{count: 0}, fn r -> r.response == &1["id"] end).count
-        })
-      )
-      |> then(fn choices ->
-        blank_reponses = Enum.find(responses, fn r -> r.response == "" end)
-
-        if blank_reponses[:response] do
-          [
-            %{
-              "content" => [
-                %{
-                  "children" => [
-                    %{
-                      "text" =>
-                        "Blank attempt (user submitted assessment without selecting any choice for this activity)"
-                    }
-                  ],
-                  "type" => "p"
-                }
-              ],
-              "frequency" => blank_reponses.count
-            }
-            | choices
-          ]
-        else
-          choices
-        end
-      end)
-
-    update_in(
-      activity_attempt,
-      [Access.key!(:revision), Access.key!(:content)],
-      &Map.put(&1, "choices", choices)
-    )
-    |> update_in(
-      [
-        Access.key!(:revision),
-        Access.key!(:content)
-      ],
-      &Map.put(&1, "activityTitle", activity_attempt.revision.title)
-    )
-  end
-
-  defp add_multi_input_details(activity_attempt, response_summaries) do
-    mapper = build_input_mapper(activity_attempt.transformed_model["inputs"])
-
-    Enum.reduce(
-      activity_attempt.transformed_model["inputs"],
-      activity_attempt,
-      fn input, acc2 ->
-        case input["inputType"] do
-          response when response in ["numeric", "text"] ->
-            add_text_or_numeric_responses(
-              acc2,
-              response_summaries,
-              mapper
-            )
-
-          "dropdown" ->
-            add_dropdown_choices(acc2, response_summaries)
-        end
-      end
-    )
-  end
-
-  defp add_dropdown_choices(acc, response_summaries) do
-    add_choices_frequencies(acc, response_summaries)
-    |> update_in(
-      [
-        Access.key!(:transformed_model),
-        Access.key!("inputs"),
-        Access.filter(&(&1["inputType"] == "dropdown")),
-        Access.key!("choiceIds")
-      ],
-      &List.insert_at(&1, -1, "0")
-    )
-  end
-
-  defp add_text_or_numeric_responses(acumulator, response_summaries, mapper) do
-    responses =
-      relevant_responses(acumulator.resource_id, response_summaries, mapper)
-
-    update_in(
-      acumulator,
-      [Access.key!(:transformed_model), Access.key!("authoring")],
-      &Map.put(&1, "responses", responses)
-    )
-  end
-
-  defp relevant_responses(resource_id, response_summaries, mapper) do
-    Enum.reduce(response_summaries, [], fn response_summary, acc_responses ->
-      if response_summary.activity_id == resource_id do
-        [
-          %{
-            text: response_summary.response,
-            user_name: OliWeb.Common.Utils.name(response_summary.user),
-            type: mapper[response_summary.part_id],
-            part_id: response_summary.part_id
-          }
-          | acc_responses
-        ]
-      else
-        acc_responses
-      end
-    end)
-  end
-
-  defp build_input_mapper(inputs) do
-    Enum.into(inputs, %{}, fn input ->
-      {input["partId"], input["inputType"]}
-    end)
   end
 end
